fix: separate reducer code
